edu/requirements.txt
google-generativeai
python-dotenv
numpy
scikit-learn